Migrate GyokuonCommand
